refactor header components and styling
