+++ src/App.js
@@ -3,7 +3,6 @@ import './App.css';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import './Assets/css/custom.css'
 import Header from './components/Header/header';
-import Navbar from './components/styles/Nav/Navbar';
 import Footer from './components/Footer/footer';
 import About from './components/About/about';
 import Portfolio from './components/Portfolio/portfolio';
@@ -35,7 +34,6 @@ function App() {
   return (
 		<div className='App custom-fonts'>
 			<Header currentPage={currentPage} handlePageChange={handlePageChange} />
-      {/* <Navbar /> */}
 			<main>
 				<RenderPage />
 			</main>
